microservices/transforming_pipeline/components/anonimization/requirements.txt
pandas
python-dotenv
openai
tqdm
pytest
streamlit
pipeline